requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via aiohttp
certifi==2025.7.9
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   yarl
jiter==0.10.0
    # via openai
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
openai==1.95.1
    # via -r requirements.in
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tenacity==9.1.2
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
yarl==1.20.1
    # via aiohttp
